requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
PyMySQL>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
redis>=5.0.0
openai>=1.0.0
